chore: rename standard to more appopriately match its use
