+++ tomb-crypt/src/key_seal/native/ec_encryption_key.rs
@@ -1,8 +1,8 @@
 use openssl::pkey::{PKey, Private};
 
-use crate::key_seal::KeySealError;
 use crate::key_seal::common::*;
-use crate::key_seal::standard::*;
+use crate::key_seal::native::*;
+use crate::key_seal::KeySealError;
 
 pub struct EcEncryptionKey(pub(crate) PKey<Private>);
