+++ tomb-crypt/src/key_seal/native/encrypted_symmetric_key.rs
@@ -1,6 +1,6 @@
-use crate::key_seal::KeySealError;
 use crate::key_seal::common::*;
-use crate::key_seal::standard::*;
+use crate::key_seal::native::*;
+use crate::key_seal::KeySealError;
 
 pub struct EncryptedSymmetricKey {
     pub(crate) data: [u8; AES_KEY_SIZE + 8],
@@ -13,10 +13,7 @@ impl ProtectedKey for EncryptedSymmetricKey {
     type PlainKey = SymmetricKey;
     type WrappingPrivateKey = EcEncryptionKey;
 
-    fn decrypt_with(
-        &self,
-        recipient_key: &EcEncryptionKey,
-    ) -> Result<SymmetricKey, KeySealError> {
+    fn decrypt_with(&self, recipient_key: &EcEncryptionKey) -> Result<SymmetricKey, KeySealError> {
         let ephemeral_public_key = EcPublicEncryptionKey::import_bytes(self.public_key.as_ref())?;
         let ecdh_shared_secret = internal::ecdh_exchange(&recipient_key.0, &ephemeral_public_key.0);
