+++ tomb-crypt/src/key_seal/native/error.rs
